Support refresh and self identification by a host
Will allow support for periodic refresh of the url backed hosts
