+++ app/src/main/java/acr/browser/lightning/adblock/source/AssetsHostsDataSource.kt
@@ -1,5 +1,6 @@
 package acr.browser.lightning.adblock.source
 
+import acr.browser.lightning.BuildConfig
 import acr.browser.lightning.adblock.HostsFileParser
 import acr.browser.lightning.log.Logger
 import android.content.res.AssetManager
@@ -37,6 +38,10 @@ class AssetsHostsDataSource @Inject constructor(
         logger.log(TAG, "Loaded ${domains.size} domains")
         emitter.onSuccess(HostsResult.Success(domains))
     }
+
+    override fun identifier(): String = "assets:${BuildConfig.VERSION_CODE}"
+
+    override fun requiresRefresh(): Boolean = false
 
     companion object {
         private const val TAG = "AssetsHostsDataSource"
