+++ app/src/main/java/acr/browser/lightning/adblock/source/HostsDataSourceProvider.kt
@@ -10,9 +10,4 @@ interface HostsDataSourceProvider {
      */
     fun createHostsDataSource(): HostsDataSource
 
-    /**
-     * The identity of the hosts data source.
-     */
-    fun sourceIdentity(): String
-
 }
